+++ zest/releaser/utils.py
@@ -260,7 +260,7 @@ def setup_py(rest_of_cmdline):
     executable = sys.executable
     if TESTMODE:
         # Hack for testing
-        for unsafe in ['upload', 'register', 'mregister']:
+        for unsafe in ['upload', 'register']:
             if unsafe in rest_of_cmdline:
                 executable = 'echo MOCK'